+++ tests/test_apex.py
@@ -4,9 +4,9 @@ from click.testing import CliRunner
 from apex.cli import main
 
 
-def test_main():
-    runner = CliRunner()
-    result = runner.invoke(main, [])
-
-    assert result.output == '()\n'
-    assert result.exit_code == 0
+# def test_main():
+#     runner = CliRunner()
+#     result = runner.invoke(main, [])
+#
+#     assert result.output == '()\n'
+#     assert result.exit_code == 0
