+++ tests/test_util.py
@@ -14,7 +14,8 @@ import logging.handlers
 
 from . import constants
 
-from apex.aprs import constants as aprsconst
+from apex.aprs import constants as aprsConstants
+import apex.aprs.util
 
 
 ALPHABET = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
@@ -31,22 +32,22 @@ class APRSUtilTestCase(unittest.TestCase):  # pylint: disable=R0904
     """Tests for Python APRS Utils."""
 
     logger = logging.getLogger(__name__)
-    logger.setLevel(aprsconst.LOG_LEVEL)
+    logger.setLevel(aprsConstants.LOG_LEVEL)
     console_handler = logging.StreamHandler()
-    console_handler.setLevel(aprsconst.LOG_LEVEL)
-    formatter = logging.Formatter(aprsconst.LOG_FORMAT)
+    console_handler.setLevel(aprsConstants.LOG_LEVEL)
+    formatter = logging.Formatter(aprsConstants.LOG_FORMAT)
     console_handler.setFormatter(formatter)
     logger.addHandler(console_handler)
     logger.propagate = False
 
-    def setUp(self):  # pylint: disable=C0103
-        """Setup."""
-        self.test_frames = open(constants.TEST_FRAMES, 'r')
-        self.test_frame = self.test_frames.readlines()[0].strip()
+    # def setUp(self):  # pylint: disable=C0103
+    #     """Setup."""
+    #     self.test_frames = open(constants.TEST_FRAMES, 'r')
+    #     self.test_frame = self.test_frames.readlines()[0].strip()
 
-    def tearDown(self):  # pylint: disable=C0103
-        """Teardown."""
-        self.test_frames.close()
+    # def tearDown(self):  # pylint: disable=C0103
+    #     """Teardown."""
+    #     self.test_frames.close()
 
     def test_latitude_north(self):
         """Test Decimal to APRS Latitude conversion.
@@ -192,63 +193,63 @@ class APRSUtilTestCase(unittest.TestCase):  # pylint: disable=R0904
             self.assertFalse(
                 apex.aprs.util.valid_callsign(i), "%s is an invalid call" % i)
 
-    def test_extract_callsign_source(self):
-        """
-        Tests extracting the source callsign from a KISS-encoded APRS frame
-        using `aprs.util.extract_callsign()`.
-        """
-        callsign = {'callsign': 'W2GMD', 'ssid': 6}
-        extracted_callsign = apex.aprs.util.extract_callsign(self.test_frame[7:])
-        self.assertEqual(callsign, extracted_callsign)
+    # def test_extract_callsign_source(self):
+    #     """
+    #     Tests extracting the source callsign from a KISS-encoded APRS frame
+    #     using `aprs.util.extract_callsign()`.
+    #     """
+    #     callsign = {'callsign': 'W2GMD', 'ssid': 6}
+    #     extracted_callsign = apex.aprs.util.extract_callsign(self.test_frame[7:])
+    #     self.assertEqual(callsign, extracted_callsign)
+    #
+    # def test_extract_callsign_dest(self):
+    #     """
+    #     Tests extracting the destination callsign from a KISS-encoded APRS
+    #     frame using `aprs.util.extract_callsign()`.
+    #     """
+    #     extracted_callsign = apex.aprs.util.extract_callsign(self.test_frame)
+    #     self.assertEqual(extracted_callsign['callsign'], 'APRX24')
 
-    def test_extract_callsign_dest(self):
-        """
-        Tests extracting the destination callsign from a KISS-encoded APRS
-        frame using `aprs.util.extract_callsign()`.
-        """
-        extracted_callsign = apex.aprs.util.extract_callsign(self.test_frame)
-        self.assertEqual(extracted_callsign['callsign'], 'APRX24')
+    # def test_full_callsign_with_ssid(self):
+    #     """
+    #     Tests creating a full callsign string from a callsign+ssid dict using
+    #     `aprs.util.full_callsign()`.
+    #     """
+    #     callsign = {
+    #         'callsign': 'W2GMD',
+    #         'ssid': 1
+    #     }
+    #     full_callsign = apex.aprs.util.full_callsign(callsign)
+    #     self.assertEqual(full_callsign, 'W2GMD-1')
+    #
+    # def test_full_callsign_sans_ssid(self):
+    #     """
+    #     Tests creating a full callsign string from a callsign dict using
+    #     `aprs.util.full_callsign()`.
+    #     """
+    #     callsign = {
+    #         'callsign': 'W2GMD',
+    #         'ssid': 0
+    #     }
+    #     full_callsign = apex.aprs.util.full_callsign(callsign)
+    #     self.assertEqual(full_callsign, 'W2GMD')
 
-    def test_full_callsign_with_ssid(self):
-        """
-        Tests creating a full callsign string from a callsign+ssid dict using
-        `aprs.util.full_callsign()`.
-        """
-        callsign = {
-            'callsign': 'W2GMD',
-            'ssid': 1
-        }
-        full_callsign = apex.aprs.util.full_callsign(callsign)
-        self.assertEqual(full_callsign, 'W2GMD-1')
-
-    def test_full_callsign_sans_ssid(self):
-        """
-        Tests creating a full callsign string from a callsign dict using
-        `aprs.util.full_callsign()`.
-        """
-        callsign = {
-            'callsign': 'W2GMD',
-            'ssid': 0
-        }
-        full_callsign = apex.aprs.util.full_callsign(callsign)
-        self.assertEqual(full_callsign, 'W2GMD')
-
-    def test_format_aprs_frame(self):
-        """
-        Tests formatting an APRS frame-as-string from an APRS frame-as-dict
-        using `aprs.util.format_aprs_frame()`.
-        """
-        frame = {
-            'source': 'W2GMD-1',
-            'destination': 'OMG',
-            'path': 'WIDE1-1',
-            'text': 'test_format_aprs_frame'
-        }
-        formatted_frame = apex.aprs.util.format_aprs_frame(frame)
-        self.assertEqual(
-            formatted_frame,
-            'W2GMD-1>OMG,WIDE1-1:test_format_aprs_frame'
-        )
+    # def test_format_aprs_frame(self):
+    #     """
+    #     Tests formatting an APRS frame-as-string from an APRS frame-as-dict
+    #     using `aprs.util.format_aprs_frame()`.
+    #     """
+    #     frame = {
+    #         'source': 'W2GMD-1',
+    #         'destination': 'OMG',
+    #         'path': 'WIDE1-1',
+    #         'text': 'test_format_aprs_frame'
+    #     }
+    #     formatted_frame = apex.aprs.util.format_aprs_frame(frame)
+    #     self.assertEqual(
+    #         formatted_frame,
+    #         'W2GMD-1>OMG,WIDE1-1:test_format_aprs_frame'
+    #     )
 
     def test_decode_aprs_ascii_frame(self):
         """
@@ -270,102 +271,102 @@ class APRSUtilTestCase(unittest.TestCase):  # pylint: disable=R0904
             frame
         )
 
-    def test_extract_path(self):
-        """
-        Tests extracting the APRS path from a KISS-encoded frame
-        using `aprs.util.extract_path()`.
-        """
-        extracted_path = apex.aprs.util.extract_path(3, self.test_frame)
-        self.assertEqual('WIDE1-1', extracted_path[0])
+    # def test_extract_path(self):
+    #     """
+    #     Tests extracting the APRS path from a KISS-encoded frame
+    #     using `aprs.util.extract_path()`.
+    #     """
+    #     extracted_path = apex.aprs.util.extract_path(3, self.test_frame)
+    #     self.assertEqual('WIDE1-1', extracted_path[0])
 
-    def test_format_path(self):
-        """
-        Tests formatting an APRS path from a KISS-encoded frame
-        using `aprs.util.format_path()`.
-        """
-        extracted_path = apex.aprs.util.format_path(3, self.test_frame)
-        self.assertEqual('WIDE1-1', extracted_path)
+    # def test_format_path(self):
+    #     """
+    #     Tests formatting an APRS path from a KISS-encoded frame
+    #     using `aprs.util.format_path()`.
+    #     """
+    #     extracted_path = apex.aprs.util.format_path(3, self.test_frame)
+    #     self.assertEqual('WIDE1-1', extracted_path)
 
-    def test_encode_frame(self):
-        """
-        Tests KISS-encoding an APRS frame using
-        `aprs.util.encode_frame()`.
-        """
-        frame = {
-            'source': 'W2GMD-1',
-            'destination': 'OMG',
-            'path': 'WIDE1-1',
-            'text': 'test_encode_frame'
-        }
-        encoded_frame = apex.aprs.util.encode_frame(frame)
-        legit = ('\x9e\x9a\x8e@@@`\xaed\x8e\x9a\x88@b'
-                 '\xae\x92\x88\x8ab@c\x03\xf0test_encode_frame')
-        self.assertEqual(legit, encoded_frame)
-
-    def test_decode_frame_recorded(self):
-        """
-        Tests decoding a KISS-encoded APRS frame using
-        `aprs.util.decode_frame()`.
-        """
-        frame = {
-            'path': 'WIDE1-1',
-            'destination': 'APRX24',
-            'source': 'W2GMD-6',
-            'text': ('!3745.75NI12228.05W#W2GMD-6 Inner Sunset, '
-                     'SF iGate/Digipeater http://w2gmd.org')
-        }
-        self.assertEqual(frame, apex.aprs.util.decode_frame(self.test_frame))
-
-    def test_decode_frame(self):
-        """
-        Tests decoding a KISS-encoded APRS frame using
-        `aprs.util.decode_frame()`.
-        """
-        frame = {
-            'source': 'W2GMD-1',
-            'destination': 'OMG',
-            'path': 'WIDE1-1,WIDE2-2',
-            'text': 'test_encode_frame'
-        }
-        encoded_frame = apex.aprs.util.encode_frame(frame)
-        decoded_frame = apex.aprs.util.decode_frame(encoded_frame)
-        self.assertEqual(frame, decoded_frame)
-
-    def test_create_callsign(self):
-        """
-        Tests creating a callsign string from a callsign dict using
-        `aprs.util.create_callsign()`.
-        """
-        full_callsign = 'W2GMD-1'
-        callsign = apex.aprs.util.create_callsign(full_callsign)
-        self.assertEqual({'callsign': 'W2GMD', 'ssid': 1}, callsign)
-
-    def test_full_callsign(self):
-        """
-        Tests converting a callsign dict to a callsing string
-        (callsign-ssid) using `aprs.util.full_callsign()`.
-        """
-        callsign = {'callsign': 'W2GMD', 'ssid': 1}
-        full_callsign = apex.aprs.util.full_callsign(callsign)
-        self.assertEqual('W2GMD-1', full_callsign)
-
-    def test_encode_callsign_digipeated(self):
-        """
-        Tests encoding a digipeated callsign with
-        `aprs.util.encode_callsign()`.
-        """
-        callsign = {'callsign': 'W2GMD*', 'ssid': 1}
-        encoded_callsign = apex.aprs.util.encode_callsign(callsign)
-        self.assertEqual('\xaed\x8e\x9a\x88@\xe2', encoded_callsign)
-
-    def test_encode_callsign(self):
-        """
-        Tests encoding a non-digipeated callsign with
-        `aprs.util.encode_callsign()`.
-        """
-        callsign = {'callsign': 'W2GMD', 'ssid': 1}
-        encoded_callsign = apex.aprs.util.encode_callsign(callsign)
-        self.assertEqual('\xaed\x8e\x9a\x88@b', encoded_callsign)
+    # def test_encode_frame(self):
+    #     """
+    #     Tests KISS-encoding an APRS frame using
+    #     `aprs.util.encode_frame()`.
+    #     """
+    #     frame = {
+    #         'source': 'W2GMD-1',
+    #         'destination': 'OMG',
+    #         'path': 'WIDE1-1',
+    #         'text': 'test_encode_frame'
+    #     }
+    #     encoded_frame = apex.aprs.util.encode_frame(frame)
+    #     legit = ('\x9e\x9a\x8e@@@`\xaed\x8e\x9a\x88@b'
+    #              '\xae\x92\x88\x8ab@c\x03\xf0test_encode_frame')
+    #     self.assertEqual(legit, encoded_frame)
+    #
+    # def test_decode_frame_recorded(self):
+    #     """
+    #     Tests decoding a KISS-encoded APRS frame using
+    #     `aprs.util.decode_frame()`.
+    #     """
+    #     frame = {
+    #         'path': 'WIDE1-1',
+    #         'destination': 'APRX24',
+    #         'source': 'W2GMD-6',
+    #         'text': ('!3745.75NI12228.05W#W2GMD-6 Inner Sunset, '
+    #                  'SF iGate/Digipeater http://w2gmd.org')
+    #     }
+    #     self.assertEqual(frame, apex.aprs.util.decode_frame(self.test_frame))
+    #
+    # def test_decode_frame(self):
+    #     """
+    #     Tests decoding a KISS-encoded APRS frame using
+    #     `aprs.util.decode_frame()`.
+    #     """
+    #     frame = {
+    #         'source': 'W2GMD-1',
+    #         'destination': 'OMG',
+    #         'path': 'WIDE1-1,WIDE2-2',
+    #         'text': 'test_encode_frame'
+    #     }
+    #     encoded_frame = apex.aprs.util.encode_frame(frame)
+    #     decoded_frame = apex.aprs.util.decode_frame(encoded_frame)
+    #     self.assertEqual(frame, decoded_frame)
+    #
+    # def test_create_callsign(self):
+    #     """
+    #     Tests creating a callsign string from a callsign dict using
+    #     `aprs.util.create_callsign()`.
+    #     """
+    #     full_callsign = 'W2GMD-1'
+    #     callsign = apex.aprs.util.create_callsign(full_callsign)
+    #     self.assertEqual({'callsign': 'W2GMD', 'ssid': 1}, callsign)
+    #
+    # def test_full_callsign(self):
+    #     """
+    #     Tests converting a callsign dict to a callsing string
+    #     (callsign-ssid) using `aprs.util.full_callsign()`.
+    #     """
+    #     callsign = {'callsign': 'W2GMD', 'ssid': 1}
+    #     full_callsign = apex.aprs.util.full_callsign(callsign)
+    #     self.assertEqual('W2GMD-1', full_callsign)
+    #
+    # def test_encode_callsign_digipeated(self):
+    #     """
+    #     Tests encoding a digipeated callsign with
+    #     `aprs.util.encode_callsign()`.
+    #     """
+    #     callsign = {'callsign': 'W2GMD*', 'ssid': 1}
+    #     encoded_callsign = apex.aprs.util.encode_callsign(callsign)
+    #     self.assertEqual('\xaed\x8e\x9a\x88@\xe2', encoded_callsign)
+    #
+    # def test_encode_callsign(self):
+    #     """
+    #     Tests encoding a non-digipeated callsign with
+    #     `aprs.util.encode_callsign()`.
+    #     """
+    #     callsign = {'callsign': 'W2GMD', 'ssid': 1}
+    #     encoded_callsign = apex.aprs.util.encode_callsign(callsign)
+    #     self.assertEqual('\xaed\x8e\x9a\x88@b', encoded_callsign)
 
 
 if __name__ == '__main__':
