+++ tox.ini
@@ -10,8 +10,8 @@ envlist =
     3.4-nocov,
     3.5-cover,
     3.5-nocov,
-    pypy-cover,
-    pypy-nocov,
+;    pypy-cover,
+;    pypy-nocov,
     report,
     docs
 
@@ -114,20 +114,20 @@ skip_install = true
 usedevelop = false
 deps = coverage
 
-[testenv:2.7-cover]
-basepython = {env:TOXPYTHON:python2.7}
-setenv =
-    {[testenv]setenv}
-    WITH_COVERAGE=yes
-usedevelop = true
-commands =
-    {posargs:py.test --cov --cov-report=term-missing -vv}
-deps =
-    {[testenv]deps}
-    pytest-cov
-
-[testenv:2.7-nocov]
-basepython = {env:TOXPYTHON:python2.7}
+;[testenv:2.7-cover]
+;basepython = {env:TOXPYTHON:python2.7}
+;setenv =
+;    {[testenv]setenv}
+;    WITH_COVERAGE=yes
+;usedevelop = true
+;commands =
+;    {posargs:py.test --cov --cov-report=term-missing -vv}
+;deps =
+;    {[testenv]deps}
+;    pytest-cov
+;
+;[testenv:2.7-nocov]
+;basepython = {env:TOXPYTHON:python2.7}
 
 [testenv:3.3-cover]
 basepython = {env:TOXPYTHON:python3.3}
@@ -174,20 +174,20 @@ deps =
 [testenv:3.5-nocov]
 basepython = {env:TOXPYTHON:python3.5}
 
-[testenv:pypy-cover]
-basepython = {env:TOXPYTHON:pypy}
-setenv =
-    {[testenv]setenv}
-    WITH_COVERAGE=yes
-usedevelop = true
-commands =
-    {posargs:py.test --cov --cov-report=term-missing -vv}
-deps =
-    {[testenv]deps}
-    pytest-cov
-
-[testenv:pypy-nocov]
-basepython = {env:TOXPYTHON:pypy}
-
-
+;[testenv:pypy-cover]
+;basepython = {env:TOXPYTHON:pypy}
+;setenv =
+;    {[testenv]setenv}
+;    WITH_COVERAGE=yes
+;usedevelop = true
+;commands =
+;    {posargs:py.test --cov --cov-report=term-missing -vv}
+;deps =
+;    {[testenv]deps}
+;    pytest-cov
+;
+;[testenv:pypy-nocov]
+;basepython = {env:TOXPYTHON:pypy}
+;
+;
